+++ src/main/java/de/sakurajin/sakuralib/compat/CompatLoader.java
@@ -1,0 +1,12 @@
+package de.sakurajin.sakuralib.compat;
+
+/**
+ * This class is used to load the compat modules.
+ * It is called from the mod initializer and simply calls the init functions of the compat modules.
+ * Its whole pupose is to make the mod initializer a bit cleaner.
+ */
+public class CompatLoader {
+    public static void init(){
+
+    }
+}