+++ src/main/java/de/sakurajin/sakuralib/internal/SakuralibConfigModel.java
@@ -10,6 +10,10 @@ public class SakuralibConfigModel {
 
     public boolean DEBUG = false;
 
+    public boolean ALWAYS_ADD_EXAMPLE_DATA = false;
+
+    public boolean REGISTER_PATCHOULI_DATA_ON_RELOAD = false;
+
     @SectionHeader("ModTracking")
     public boolean ENABLE_WARNING_SCREEN = false;
     public boolean WARN_ON_MAJOR_VERSION_UPDATE = false;